requirements.txt
# Core training stack (CPU wheels by default)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Ultralytics YOLO (v11/v8 interface)
ultralytics==8.3.221

# Common deps used by Ultralytics and training utilities
opencv-python==4.12.0.88
matplotlib==3.10.7
numpy==2.2.6
pillow==11.3.0
pyyaml==6.0.3
requests==2.32.5
scipy==1.16.2
psutil==7.1.2
polars==1.34.0

# Note for CUDA users (optional):
# If you have an NVIDIA GPU and want CUDA-enabled PyTorch, install with:
#   pip install --index-url https://download.pytorch.org/whl/cu121 torch torchvision torchaudio

# Web framework and dependencies
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Database support
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
pymysql==1.1.0  # MySQL driver (alternative)

# Environment variables
python-dotenv==1.0.0

